microservices/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.3
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
requests==2.31.0
fyers-apiv3==3.1.7
httpx==0.25.2
bcrypt==4.1.2
PyJWT==2.8.0
psutil==5.9.6

